Add crossbow plugins example and bump macroquad and bevy versions
