+++ examples/bevy-2d/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 crossbow = { version = "0.2.0", path = "../../" }
 log = "0.4"
 anyhow = "1.0"
-bevy = { version = "0.7.0", features = ["mp3"] }
+bevy = { version = "0.8.0", features = ["mp3"] }
 
 [package.metadata]
 app_name = "Bevy 2D"
