+++ examples/bevy-2d/src/main.rs
@@ -11,7 +11,7 @@ fn main() {
 }
 
 fn icon(mut commands: Commands, asset_server: Res<AssetServer>) {
-    commands.spawn_bundle(OrthographicCameraBundle::new_2d());
+    commands.spawn_bundle(Camera2dBundle::default());
     commands.spawn_bundle(SpriteBundle {
         texture: asset_server.load("images/icon.png"),
         ..Default::default()
