+++ examples/bevy-explorer/Cargo.toml
@@ -10,7 +10,7 @@ log = "0.4"
 anyhow = "1.0"
 subxt = "0.21"
 tokio = { version = "1.17", features = ["sync", "macros", "rt-multi-thread"] }
-bevy = { version = "0.7.0", features = ["mp3"] }
+bevy = { version = "0.8.0", features = ["mp3"] }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive", "full", "bit-vec"] }
 jsonrpsee = { version = "0.10.1", features = ["async-client", "client-ws-transport"] }
