+++ examples/crossbow-plugins/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "macroquad-play-billing"
+name = "crossbow-plugins"
 version = "0.2.0"
 authors = ["DodoRare Team <support@dodorare.com>"]
 edition = "2021"
@@ -8,13 +8,14 @@ edition = "2021"
 crossbow = { version = "0.2.0", path = "../../" }
 log = "0.4"
 anyhow = "1.0"
-macroquad = "0.3.7"
+macroquad = { version = "0.3", default-features = false }
 
 [target.'cfg(target_os = "android")'.dependencies]
 play-billing = { version = "0.2.0", path = "../../plugins/play-billing" }
+play-games-services = { version = "0.2.0", path = "../../plugins/play-games-services" }
 
 [package.metadata]
-app_name = "Play Billing"
+app_name = "Crossbow Plugins"
 
 [package.metadata.android]
 app_wrapper = "sokol"
@@ -30,12 +31,18 @@ include = ":crossbow:lib"
 [[package.metadata.android.plugins_local_projects]]
 include = ":play_billing"
 project_dir = "../../plugins/play-billing/android"
+[[package.metadata.android.plugins_local_projects]]
+include = ":play_games_services"
+project_dir = "../../plugins/play-games-services/android"
 
 [package.metadata.android.manifest]
-package = "com.crossbow.example.play_billing"
+package = "com.crossbow.example.plugins"
 [package.metadata.android.manifest.uses_sdk]
 min_sdk_version = 19
 target_sdk_version = 30
-[package.metadata.android.manifest.application]
-icon = "@mipmap/ic_launcher"
-round_icon = "@mipmap/ic_launcher"
+[[package.metadata.android.manifest.application.meta_data]]
+name = "com.google.android.gms.games.APP_ID"
+value = "@string/app_id"
+[[package.metadata.android.manifest.application.meta_data]]
+name = "com.google.android.gms.version"
+value = "@integer/google_play_services_version"
