+++ examples/macroquad-3d/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 crossbow = { version = "0.2.0", path = "../../" }
 log = "0.4"
 anyhow = "1.0"
-macroquad = "0.3.7"
+macroquad = { version = "0.3", default-features = false }
 
 [package.metadata]
 app_name = "Macroquad 3D"
