+++ examples/macroquad-permissions/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 crossbow = { version = "0.2.0", path = "../../" }
 log = "0.4"
 anyhow = "1.0"
-macroquad = "0.3.7"
+macroquad = { version = "0.3", default-features = false }
 
 [target.'cfg(target_os = "android")'.dependencies]
 admob-android = { version = "0.2.0", path = "../../plugins/admob-android" }
@@ -37,14 +37,6 @@ package = "com.crossbow.example.permissions"
 [package.metadata.android.manifest.uses_sdk]
 min_sdk_version = 19
 target_sdk_version = 30
-
-[package.metadata.android.manifest.application]
-icon = "@mipmap/ic_launcher"
-round_icon = "@mipmap/ic_launcher"
-# [[package.metadata.android.manifest.application.activity]]
-# name = ""
-# config_changes = "screenSize|orientation"
-# screen_orientation = "landscape"
 
 [package.metadata.apple]
 release_build_targets = ["aarch64-apple-ios", "x86_64-apple-ios"]
