--- examples/macroquad-play-games/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "macroquad-play-games"
-version = "0.2.0"
-authors = ["DodoRare Team <support@dodorare.com>"]
-edition = "2021"
-
-[dependencies]
-crossbow = { version = "0.2.0", path = "../../" }
-log = "0.4"
-anyhow = "1.0"
-macroquad = "0.3.7"
-
-[target.'cfg(target_os = "android")'.dependencies]
-play-games-services = { version = "0.2.0", path = "../../plugins/play-games-services" }
-
-[package.metadata]
-app_name = "Play Games"
-
-[package.metadata.android]
-app_wrapper = "sokol"
-release_build_targets = ["aarch64-linux-android"]
-resources = ["../../assets/res/android"]
-
-[[package.metadata.android.plugins_local_projects]]
-include = ":crossbow"
-dont_implement = true
-project_dir = "../../platform/android/java"
-[[package.metadata.android.plugins_local_projects]]
-include = ":crossbow:lib"
-[[package.metadata.android.plugins_local_projects]]
-include = ":play_games_services"
-project_dir = "../../plugins/play-games-services/android"
-
-[package.metadata.android.manifest]
-package = "com.crossbow.example.play_games"
-[package.metadata.android.manifest.uses_sdk]
-min_sdk_version = 19
-target_sdk_version = 30
-[package.metadata.android.manifest.application]
-icon = "@mipmap/ic_launcher"
-round_icon = "@mipmap/ic_launcher"
-[[package.metadata.android.manifest.application.meta_data]]
-name = "com.google.android.gms.games.APP_ID"
-value = "@string/app_id"
-[[package.metadata.android.manifest.application.meta_data]]
-name = "com.google.android.gms.version"
-value = "@integer/google_play_services_version"
