+++ plugins/admob-android/android/build.gradle
@@ -25,7 +25,7 @@ repositories {
 dependencies {
     implementation libraries.kotlinStdLib
     implementation libraries.androidxAppcompat
-    implementation "com.google.android.gms:play-services-ads:20.4.0"
+    implementation "com.google.android.gms:play-services-ads:21.1.0"
     implementation "com.google.android.ump:user-messaging-platform:2.0.0"
 
     if (rootProject.findProject(":crossbow:lib")) {
